delete old Courses component
