+++ frontend/src/containers/Courses/CourseCreateModal/CourseCreateModal.styles.ts
@@ -1,20 +1,14 @@
 import {createStyles, Theme} from "@material-ui/core";
 
 export default (theme: Theme) => createStyles({
-    input: {
-        width: '550px',
-        marginBottom: '30px',
-    },
     lastInput: {
         width: '550px',
-    },
-    selector: {
-        width: '550px'
     },
     actions: {
         padding: '15px 24px 20px'
     },
     dialog: {
+        width: '900px',
         padding: 20,
     },
 });
