+++ frontend/src/containers/Courses/CourseCreateModal/FirstStep/FirstStep.styles.ts
