+++ frontend/src/containers/Courses/CourseCreateModal/SecondStep/SecondStep.styles.ts
