+++ frontend/src/containers/Courses/CourseCreateModal/SecondStep/SecondStep.tsx
