+++ frontend/src/containers/Courses/Courses.styles.ts
@@ -48,9 +48,19 @@ export default (theme: Theme) => createStyles({
         marginLeft: 'auto',
         padding: '0px 20px'
     },
+    searchWrapper: {},
     searchInput: {
         height: '40px',
         minHeight: '40px'
+    },
+    filterBtn: {
+        width: '150px',
+        height: '40px',
+        marginRight: '20px',
+        backgroundColor: theme.palette.primary.main,
+        color: '#fff',
+        textDecoration: 'uppercase',
+        fontSize: '15px',
     },
     addIcon: {
         marginLeft: 'auto',
@@ -61,4 +71,8 @@ export default (theme: Theme) => createStyles({
         alignItems: 'center',
         marginTop: '10px'
     },
+    accordionSummary: {
+        backgroundColor: 'rgba(0, 0, 0, .03)', 
+        borderBottom: '1px solid rgba(0, 0, 0, .125)',
+    }
 });
