+++ frontend/src/containers/Courses/CoursesTable/CoursesTable.styles.ts
