+++ frontend/src/containers/Courses/actions.ts
@@ -5,12 +5,19 @@ import {CoursesActions} from './types';
 const getCourses = createAction('GET_COURSES');
 const setCourses = createAction('SET_COURSES');
 
+const getPlatforms = createAction('GET_PLATFORMS')
+const setPlatforms = createAction('SET_PLATFORMS')
+
+const getInstitutions = createAction('GET_INSTITUTIONS')
+const setInstitutions = createAction('SET_INSTITUTIONS')
+
 const createNewCourse = createAction('CREATE_NEW_COURSE');
 const changeCourse = createAction('CHANGE_COURSE');
 const deleteCourse = createAction('DELETE_COURSE');
 
 const openDialog = createAction('OPEN_COURSE_DIALOG');
 const closeDialog = createAction('CLOSE_COURSE_DIALOG');
+const changeDialogStep = createAction('CHANGE_DIALOG_STEP')
 
 const changeSearchQuery = createAction('COURSES_CHANGE_SEARCH_QUERY');
 const changeCurrentPage = createAction('COURSES_CHANGE_CURRENT_PAGE');
@@ -29,6 +36,11 @@ const actions: CoursesActions = {
     changeCurrentPage,
     changeAllCount,
     changeSorting,
+    changeDialogStep,
+    getPlatforms,
+    setPlatforms,
+    getInstitutions,
+    setInstitutions
 }
 
 export default actions;
