+++ frontend/src/containers/Courses/logics.ts
@@ -42,81 +42,44 @@ const getCourses = createLogic({
             });
     }
 });
-
-const deleteCourse = createLogic({
-    type: courseActions.deleteCourse.type,
+// ONLY 1st PAGE
+const getPlatforms = createLogic({
+    type: courseActions.getPlatforms.type,
     latest: true,
     process({getState, action}: any, dispatch, done) {
-        const courseId = action.payload;
-
-        dispatch(actions.fetchingTrue({destination: fetchingTypes.DELETE_COURSE}));
-
-        service.deleteCourse(courseId)
-            .then((res) => {
-                dispatch(courseActions.getCourses());
-                dispatch(actions.fetchingSuccess());
-            })
-            .catch((err) => {
-                dispatch(actions.fetchingFailed(err));
-            })
-            .then(() => {
-                dispatch(actions.fetchingFalse({destination: fetchingTypes.DELETE_COURSE}));
-                return done();
-            });
+        dispatch(actions.fetchingTrue({destination: fetchingTypes.GET_PLATFORMS}));
+        service.getPlatforms().then((res) => {
+            const platforms = get(res, 'data.results', [])
+            dispatch(courseActions.setPlatforms(platforms))
+        }).catch((err) => {
+            dispatch(actions.fetchingFailed(err))
+        }).then(() => {
+            dispatch(actions.fetchingFalse({destination: fetchingTypes.GET_PLATFORMS}));
+            return done();
+        })
     }
-});
+})
 
-const createNewCourse = createLogic({
-    type: courseActions.createNewCourse.type,
+// ONLY 1st PAGE
+const getInstitutions = createLogic({
+    type: courseActions.getPlatforms.type,
     latest: true,
     process({getState, action}: any, dispatch, done) {
-        const course = action.payload;
-
-        dispatch(actions.fetchingTrue({destination: fetchingTypes.CREATE_COURSE}));
-
-        service.createCourse(course)
-            .then((res) => {
-                dispatch(courseActions.getCourses());
-                dispatch(actions.fetchingSuccess());
-                dispatch(courseActions.closeDialog());
-            })
-            .catch((err) => {
-                dispatch(actions.fetchingFailed(err));
-            })
-            .then(() => {
-                dispatch(actions.fetchingFalse({destination: fetchingTypes.CREATE_COURSE}));
-                return done();
-            });
+        dispatch(actions.fetchingTrue({destination: fetchingTypes.GET_INSTITUTIONS}));
+        service.getInstitutions().then((res) => {
+            const institutions = get(res, 'data.results', [])
+            dispatch(courseActions.setInstitutions(institutions))
+        }).catch((err) => {
+            dispatch(actions.fetchingFailed(err))
+        }).then(() => {
+            dispatch(actions.fetchingFalse({destination: fetchingTypes.GET_INSTITUTIONS}));
+            return done();
+        })
     }
-});
-
-const updateCourse = createLogic({
-    type: courseActions.changeCourse.type,
-    latest: true,
-    process({getState, action}: any, dispatch, done) {
-        const course = action.payload;
-
-        dispatch(actions.fetchingTrue({destination: fetchingTypes.UPDATE_COURSE}));
-
-        service.updateCourse(course)
-            .then((res) => {
-                dispatch(courseActions.getCourses());
-                dispatch(actions.fetchingSuccess());
-                dispatch(courseActions.closeDialog());
-            })
-            .catch((err) => {
-                dispatch(actions.fetchingFailed(err));
-            })
-            .then(() => {
-                dispatch(actions.fetchingFalse({destination: fetchingTypes.UPDATE_COURSE}));
-                return done();
-            });
-    }
-});
+})
 
 export default [
     getCourses,
-    deleteCourse,
-    createNewCourse,
-    updateCourse,
+    getPlatforms,
+    getInstitutions,
 ];
