--- frontend/src/containers/OnlineCourses/CourseCreateModal/CourseCreateModal.connect.ts
@@ -1,23 +1,0 @@
-import {Dispatch} from "react";
-import {connect} from 'react-redux';
-import {bindActionCreators} from "redux";
-
-import actions from "../actions";
-import {isOpenDialog} from '../getters';
-import {CoursesActions} from "../types";
-
-import {rootState} from "../../../store/reducers";
-
-const mapStateToProps = (state: rootState) => {
-    return {
-        isOpen: isOpenDialog(state),
-    };
-};
-
-const mapDispatchToProps = (dispatch: Dispatch<CoursesActions>) => ({
-    // @ts-ignore
-    actions: bindActionCreators(actions, dispatch),
-});
-
-// @ts-ignore
-export default connect(mapStateToProps, mapDispatchToProps);
