--- frontend/src/containers/OnlineCourses/CourseCreateModal/CourseCreateModal.styles.ts
@@ -1,14 +1,0 @@
-import {createStyles, Theme} from "@material-ui/core";
-
-export default (theme: Theme) => createStyles({
-    lastInput: {
-        width: '550px',
-    },
-    actions: {
-        padding: '15px 24px 20px'
-    },
-    dialog: {
-        width: '900px',
-        padding: 20,
-    },
-});
