--- frontend/src/containers/OnlineCourses/CourseCreateModal/index.ts
@@ -1,1 +1,0 @@
-export {default} from './CourseCreateModal';
