--- frontend/src/containers/OnlineCourses/CourseCreateModal/types.d.ts
@@ -1,10 +1,0 @@
-import {WithStyles} from "@material-ui/core";
-import {CoursesActions} from '../types';
-
-import styles from "./CourseCreateModal.styles";
-
-export interface CourseCreateModalProps extends WithStyles<typeof styles> {
-    actions: CoursesActions;
-    isOpen: boolean;
-    step: number;
-}
