--- frontend/src/containers/OnlineCourses/index.ts
@@ -1,1 +1,0 @@
-export {default} from './OnlineCourses';
