--- frontend/src/containers/OnlineCourses/service.ts
@@ -1,20 +1,0 @@
-import AnalyticsService from "../../service/analytics-service";
-import {CourseFields} from "./enum";
-import {SortingType, Types} from "../../components/SortingButton/types";
-
-class CoursesServices extends AnalyticsService{
-    getCourses(currentPage: number, searchQuery: string, sortingField: string, sortingMode: SortingType){
-        const sortingSymbol = sortingMode === Types.ASC ? '-' : sortingMode === Types.DESC ? '+' : '';
-
-        //return this.get(`/api/onlinecourse?page=${currentPage}&search=${searchQuery}&ordering=${sortingSymbol}${sortingField}`);
-        return this.get(`/api/course/onlinecourse/?page=${currentPage}&search=${searchQuery}&ordering=${sortingSymbol}${sortingField}`)
-    }
-    getPlatforms(){
-        return this.get('/api/course/platform/')
-    }
-    getInstitutions(){
-        return this.get('/api/course/institution/')
-    }
-}
-
-export default CoursesServices;
