+++ frontend/src/routes.tsx
@@ -7,7 +7,6 @@ import SignIn from './containers/SignIn';
 import SignUp from './containers/SignUp';
 import WorkProgram from './containers/WorkProgram';
 import Courses from './containers/Courses';
-import OnlineCourses from './containers/OnlineCourses';
 import { Course } from './containers/Course/Course';
 import SubjectArea from './containers/SubjectArea';
 import Prerequisites from './containers/TrainingEntities';
@@ -85,11 +84,8 @@ export default () => (
             <AuthRoute path={routerService.getFoldersRoute()}>
                 <Folders />
             </AuthRoute>
-            {/* <AuthRoute path={routerService.getCoursesRoute()}>
+            <AuthRoute path={routerService.getCoursesRoute()}>
                 <Courses />
-            </AuthRoute> */}
-            <AuthRoute path={routerService.getOnlineCoursesRoute()}>
-                <OnlineCourses />
             </AuthRoute>
             <AuthRoute path={routerService.getCourseRoute()}>
                 <Course />
