+++ frontend/src/service/router-service.ts
@@ -4,7 +4,6 @@ const FOLDERS = 'folders';
 const SIGN_IN = 'sign-in';
 const SIGN_UP = 'sign-up';
 const COURSES = 'courses';
-const ONLINE_COURSES = 'online-courses';
 const COURSE = 'course'
 const LITERATURE = 'literature';
 const SUBJECT_AREA = 'subject-area';
@@ -56,10 +55,6 @@ export default class RouterService {
 
     getCoursesRoute = () => {
         return SEPARATOR + COURSES;
-    };
-
-    getOnlineCoursesRoute = () => {
-        return SEPARATOR + ONLINE_COURSES;
     };
 
     getCourseRoute = () => {
