+++ frontend/src/store/logics.ts
@@ -1,8 +1,7 @@
 import SignInLogic from '../containers/SignIn/logics';
 import SignUpLogic from '../containers/SignUp/logics';
 import WorkProgramLogic from '../containers/WorkProgram/logics';
-//import CoursesLogic from '../containers/Courses/logics';
-import OnlineCoursesLogic from '../containers/OnlineCourses/logics';
+import CoursesLogic from '../containers/Courses/logics';
 import CourseLogic from '../containers/Course/logics';
 import LiteratureLogic from '../containers/Literature/logics';
 import SubjectAreaLogic from '../containers/SubjectArea/logics';
@@ -37,8 +36,7 @@ export default [
     ...SignInLogic,
     ...SignUpLogic,
     ...WorkProgramLogic,
-    //...CoursesLogic,
-    ...OnlineCoursesLogic,
+    ...CoursesLogic,
     ...LiteratureLogic,
     ...SubjectAreaLogic,
     ...PrerequisitesLogic,
